Move tests of new_pattern in its dedicated file
